scripts_old_pipeline/requirements.txt
openai-whisper
torch
tqdm
requests
numpy